pyproject.toml
[tool.poetry]
name = "therapy-talks"
version = "0.1.0"
description = ""
authors = ["greyxray <greyxray@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
streamlit-app = "^0.0.3"
openai = "0.28.0"
# openai = "^1.52.0"
loguru = "^0.7.2"
pandas = "^2.2.3"
plotly = "^5.24.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
